#9694: Cannot delete calender
https://forge.tine20.org/mantisbt/view.php?id=9694

Reviewed-on: http://gerrit.tine20.com/customers/512
Tested-by: Jenkins CI (http://ci.tine20.com/)

+++ tine20/Tinebase/Controller/Record/Abstract.php
@@ -1407,7 +1407,7 @@ abstract class Tinebase_Controller_Record_Abstract
                 }
             }
         }
-        if ($_record->has('attachments')) {
+        if ($_record->has('attachments') && Setup_Controller::getInstance()->isFilesystemAvailable()) {
             Tinebase_FileSystem_RecordAttachments::getInstance()->deleteRecordAttachments($_record);
         }
     }
